moved config interfaces
