+++ core-modular/src/core-api/global-search.ts
@@ -1,26 +1,7 @@
 import { GlobalSearchService } from '../services/global-search.service';
 import { serviceRegistry } from '../services/service-registry';
+import type { SearchResultItem } from '../types/global-search';
 import type { Luigi } from './luigi';
-
-export interface GlobalSearchProvider {
-  customSearchResultItemRenderer?: (searchResultItem: any, slot: HTMLLIElement, searchApiObj: any) => object;
-  customSearchResultRenderer?: (searchResults: any[], slot: HTMLDivElement, searchApiObj: any) => object;
-  disableInputHandlers?: boolean;
-  inputPlaceholder?: any;
-  onEnter?: () => void;
-  onEscape?: () => void;
-  onInput?: () => void;
-  onSearchBtnClick?: () => void;
-  onSearchResultItemSelected?: (searchResultItem: any) => void;
-  searchFieldCentered?: boolean;
-  toggleSearch?: (element: HTMLInputElement, visible: boolean) => void;
-}
-
-export interface SearchResultItem {
-  description: string;
-  label: string;
-  pathObject: Record<string, any>;
-}
 
 export class GlobalSearch {
   luigi: Luigi;
